+++ networking/p2p/src/main/java/tech/pegasys/teku/networking/p2p/discovery/discv5/DiscV5Service.java
@@ -24,6 +24,7 @@ import java.util.stream.Stream;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.tuweni.bytes.Bytes;
+import org.apache.tuweni.crypto.SECP256K1.SecretKey;
 import org.apache.tuweni.units.bigints.UInt64;
 import org.ethereum.beacon.discovery.DiscoverySystem;
 import org.ethereum.beacon.discovery.DiscoverySystemBuilder;
@@ -58,7 +59,7 @@ public class DiscV5Service extends Service implements DiscoveryService {
   }
 
   private final AsyncRunner asyncRunner;
-  private final Bytes localNodePrivateKey;
+  private final SecretKey localNodePrivateKey;
   private final SchemaDefinitionsSupplier currentSchemaDefinitionsSupplier;
   private final NodeRecordConverter nodeRecordConverter;
 
@@ -78,7 +79,7 @@ public class DiscV5Service extends Service implements DiscoveryService {
       final DiscoverySystemBuilder discoverySystemBuilder,
       final NodeRecordConverter nodeRecordConverter) {
     this.asyncRunner = asyncRunner;
-    this.localNodePrivateKey = privateKey;
+    this.localNodePrivateKey = SecretKeyParser.fromLibP2pPrivKey(privateKey);
     this.currentSchemaDefinitionsSupplier = currentSchemaDefinitionsSupplier;
     this.nodeRecordConverter = nodeRecordConverter;
     final String listenAddress = p2pConfig.getNetworkInterface();
@@ -95,7 +96,7 @@ public class DiscV5Service extends Service implements DiscoveryService {
             .map(NodeRecordFactory.DEFAULT::fromEnr)
             .collect(toList());
     final NodeRecordBuilder nodeRecordBuilder =
-        new NodeRecordBuilder().privateKey(privateKey).seq(seqNo);
+        new NodeRecordBuilder().secretKey(localNodePrivateKey).seq(seqNo);
     if (p2pConfig.hasUserExplicitlySetAdvertisedIp()) {
       nodeRecordBuilder.address(advertisedAddress, advertisedUdpPort, advertisedTcpPort);
     }
@@ -103,7 +104,7 @@ public class DiscV5Service extends Service implements DiscoveryService {
     this.discoverySystem =
         discoverySystemBuilder
             .listen(listenAddress, listenUdpPort)
-            .privateKey(privateKey)
+            .secretKey(localNodePrivateKey)
             .bootnodes(bootnodes)
             .localNodeRecord(localNodeRecord)
             .newAddressHandler(maybeUpdateNodeRecordHandler)