+++ CHANGELOG.md
@@ -23,6 +23,8 @@ For information on changes in released versions of Teku, see the [releases page]
   (Refer to https://github.com/ConsenSys/teku/pull/6325#issue-1409631918 for more information).  
 
 ### Additions and Improvements
+ - Improved discv5 compliance
+ - Changed the builder `is online\is offline` logs to `is available\is not available`
 
 ### Bug Fixes
  - Fix missing status filters (active, pending, exited, withdrawal) for the `/eth/v1/beacon/states/{state_id}/validators` endpoint